+++ packages/manager/src/features/ObjectStorage/BucketDetail/ObjectDetailsDrawer.test.tsx
@@ -1,6 +1,5 @@
-import { fireEvent, screen, waitFor } from '@testing-library/react';
+import { screen, waitFor } from '@testing-library/react';
 import * as React from 'react';
-import * as objectStorage from '@linode/api-v4/lib/object-storage/objects';
 import { renderWithTheme } from 'src/utilities/testHelpers';
 import ObjectDetailsDrawer, { Props } from './ObjectDetailsDrawer';
 
@@ -13,11 +12,6 @@ jest.mock('@linode/api-v4/lib/object-storage/objects', () => {
     updateObjectACL: jest.fn().mockResolvedValue({})
   };
 });
-
-const mockUpdateObjectACL = jest.spyOn<any, any>(
-  objectStorage,
-  'updateObjectACL'
-);
 
 jest.mock('src/components/EnhancedSelect/Select');
 
@@ -57,41 +51,5 @@ describe('ObjectDetailsDrawer', () => {
     await waitFor(() =>
       expect(screen.queryByTestId('lastModified')).not.toBeInTheDocument()
     );
-  });
-
-  it("shows the Object's ACL", async () => {
-    renderWithTheme(<ObjectDetailsDrawer {...props} />);
-    await waitFor(() => {
-      expect(screen.getByText('Public Read')).toHaveAttribute(
-        'aria-selected',
-        'true'
-      );
-    });
-  });
-
-  it('updates the ACL and submits the appropriate value', async () => {
-    renderWithTheme(<ObjectDetailsDrawer {...props} />);
-
-    await waitFor(() => {
-      const aclSelect = screen.getByTestId('select');
-      fireEvent.change(aclSelect, {
-        target: { value: 'private' }
-      });
-
-      expect(screen.getByText('Private')).toHaveAttribute(
-        'aria-selected',
-        'true'
-      );
-
-      const saveButton = screen.getByText('Save');
-      fireEvent.click(saveButton);
-
-      expect(mockUpdateObjectACL).toHaveBeenCalledWith(
-        'cluster-id',
-        'my-bucket',
-        'my-dir/my-image.png',
-        'private'
-      );
-    });
   });
 });
